Created skeleton code!
So, I basically all the base files that were mentioned in the write up. Basically all the things that are mandatory for us to submit in the final project. We can make our own helper functions and files as we go along, but this is just a skeleton for us to start with. I also added a Makefile and the README.txt that we should try to use as the official documentation as we go along. So, anything we write that requires documentation, we should include here.