Fixed Makefile again. memgrid and memgrind errors.
Everything now uses TARGET, so it should not be an issue. For future use, we should probably use VARIABLES for all files included in the project so that in case we change the names of one of the files, it should still not be an issue. This is simply for safety's sake. But for now, it should be fine, lol.