This is the beginning to the malloc definition. It is not yet complete.
The malloc definition is almost done. There are a few things that need to be sorted out. Right now, current issues include data not being put in the right place in memory. Other than that and some edge case checking, this method has not been so bad to write thus far...