Fixed the segmentation fault malloc was throwing when it was called twice.
So, I basically implemented all of malloc fairly easily. It now works with multiple calls and it works as expected. Memory is passed back to the caller for them to use as they wish. If you look in , you'll see that I tested it with several  values. It should be noted that the error checking here is far from complete, but the overall implementation is complete and is ready to get pushed to the  branch.